+++ src/containerapp/azext_containerapp/_help.py
@@ -2226,7 +2226,7 @@ helps['containerapp env http-route-config update'] = """
     type: command
     short-summary: Update a http route config.
     examples:
-    - name: Updates a route in the environment from a yaml file.
+    - name: Update a route in the environment from a yaml file.
       text: |
           az containerapp env http-route-config update -g MyResourceGroup -n MyEnvironment -r configname --yaml config.yaml
 """
@@ -2235,7 +2235,7 @@ helps['containerapp env http-route-config show'] = """
     type: command
     short-summary: Show a http route config.
     examples:
-    - name: Shows a route from the environment.
+    - name: Show a route from the environment.
       text: |
           az containerapp env http-route-config show -g MyResourceGroup -n MyEnvironment -r configname
 """
@@ -2244,7 +2244,7 @@ helps['containerapp env http-route-config delete'] = """
     type: command
     short-summary: Delete a http route config.
     examples:
-    - name: Deletes a route from the environment.
+    - name: Delete a route from the environment.
       text: |
           az containerapp env http-route-config delete -g MyResourceGroup -n MyEnvironment -r configname
 """